HW 4
When using a fixed url address I could parse the data and get
everything I needed. When trying to make the address a variable, I
could not parse anything. I used the same code that worked with the
fixed urls and made an attempt at what the code should be for the url
with the variable address.